refactor(meta): complete SchemaApi trait decomposition
Complete the final phase of SchemaApi refactoring by moving the
last remaining methods to their appropriate domain-specific traits.
This achieves the goal of pure trait composition where SchemaApi
contains only trait bounds with no method implementations.

Changes:
- Move `set_table_lvt()` to TableApi for table time utilities
- Move `remove_marked_deleted_index_ids()` and
  `remove_marked_deleted_table_indexes()` to IndexApi for cleanup
- Transform SchemaApi to pure trait composition pattern
- Replace generic `get()` calls with direct `get_pb()` usage
- Remove unused imports and eliminate wrapper methods

The refactoring successfully extracts all 58 methods from the
monolithic SchemaApi while maintaining backward compatibility
through trait bounds. Each domain-specific trait now handles
its own methods, improving maintainability and enabling focused
testing and development.
