+++ src/meta/api/src/schema_api_test_suite.rs
@@ -1452,25 +1452,25 @@ impl SchemaApiTestSuite {
 
             let lvt_name_ident = LeastVisibleTimeIdent::new(&tenant, table_id);
 
-            let res = mt.get(&lvt_name_ident).await?;
+            let res = mt.get_pb(&lvt_name_ident).await?;
             assert!(res.is_none());
 
             let res = mt.set_table_lvt(&lvt_name_ident, &lvt_big).await?;
             assert_eq!(res.time, time_big);
-            let res = mt.get(&lvt_name_ident).await?;
-            assert_eq!(res.unwrap().time, time_big);
+            let res = mt.get_pb(&lvt_name_ident).await?;
+            assert_eq!(res.unwrap().data.time, time_big);
 
             // test lvt never fall back
 
             let res = mt.set_table_lvt(&lvt_name_ident, &lvt_small).await?;
             assert_eq!(res.time, time_big);
-            let res = mt.get(&lvt_name_ident).await?;
-            assert_eq!(res.unwrap().time, time_big);
+            let res = mt.get_pb(&lvt_name_ident).await?;
+            assert_eq!(res.unwrap().data.time, time_big);
 
             let res = mt.set_table_lvt(&lvt_name_ident, &lvt_bigger).await?;
             assert_eq!(res.time, time_bigger);
-            let res = mt.get(&lvt_name_ident).await?;
-            assert_eq!(res.unwrap().time, time_bigger);
+            let res = mt.get_pb(&lvt_name_ident).await?;
+            assert_eq!(res.unwrap().data.time, time_bigger);
         }
 
         Ok(())
